requirements.txt
requests
pandas
openpyxl
beautifulsoup4
lxml
uvicorn
fastapi
langchain
langchain-core
langchain-google-genai
langchain-experimental
dotenv
tabulate
logging
google.genai
faiss-cpu
google-genai
tqdm